Fix: Refactoring
