+++ src/index.js
@@ -1,5 +1,5 @@
 const { ApolloServer } = require('apollo-server');
-const typeDefs = require('./schema');
+const typeDefs = require('./schema/schema');
 const resolvers = require('./resolvers');
 
 const server = new ApolloServer({
